doc-analyzer/model-experiment/requirements.txt
omegaconf
pytorch-lightning
lightning
colorlog
rootutils
boto3
python-dotenv
label-studio-sdk
tqdm
seaborn
torchvision
mlflow